voiceCommandCheck.ts init
